src/test/java/com/develogical/QueryProcessorTest.java
package com.develogical;

import jdk.nashorn.internal.runtime.ECMAException;
import org.junit.Test;

import static org.hamcrest.CoreMatchers.not;
import static org.hamcrest.CoreMatchers.containsString;
import static org.hamcrest.CoreMatchers.is;
import static org.junit.Assert.assertThat;

public class QueryProcessorTest {

    QueryProcessor queryProcessor = new QueryProcessor();

    @Test
    public void returnsEmptyStringIfCannotProcessQuery() throws Exception {
        assertThat(queryProcessor.process("test"), is(""));
    }

    @Test
    public void isNotCaseSensitive() throws Exception {
        assertThat(queryProcessor.process("shakespeare"), containsString("playwright"));
    }

    @Test
    public void knowsAboutShakespeare() throws Exception {
        assertThat(queryProcessor.process("Shakespeare"), containsString("playwright"));
    }

    @Test
    public void knowsAboutEmptyString() throws Exception {
        assertThat(queryProcessor.process(""), not(""));
    }

    @Test
    public void knowsAboutTheEmporor() throws Exception {
        assertThat(queryProcessor.process("Remi is a God"), containsString("who I am"));
    }
}
